Attach query ops moved to resource utils
